requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
Pillow
imutils